progress on using impl assets for evm template
