+++ generic-template/runtime/src/configs/mod.rs
@@ -104,6 +104,7 @@ impl AssetsConfig for OpenZeppelinConfig {
     type ApprovalDeposit = ApprovalDeposit;
     type AssetAccountDeposit = AssetAccountDeposit;
     type AssetDeposit = AssetDeposit;
+    type AssetId = u32;
     type CreateOrigin = AsEnsureOriginWithArg<EnsureSigned<AccountId>>;
     type ForceOrigin = EnsureRoot<AccountId>;
 }
